requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
requests==2.31.0
